# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Image Manipulation API",
    "description": "A simple API for image manipulation",
    "version": "1.0.0"
  },
  "scheme": [
    "http"
  ],
  "consumes": [
    "multipart/form-data"
  ],
  "produces": [
    "image/jpeg"
  ],
  "paths": {
    "/rotate": {
      "post": {
        "summary": "Rotate an image",
        "description": "Rotate an image by a specified angle",
        "tags": [
          "image"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "description": "Image file",
            "required": true,
            "type": "file"
          },
          {
            "name": "angle",
            "in": "query",
            "description": "Rotation angle (90, 180, or 270)",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/resize": {
      "post": {
        "summary": "Resize an image",
        "description": "Resize an image to a specified width and height",
        "tags": [
          "image"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "description": "Image file",
            "required": true,
            "type": "file"
          },
          {
            "name": "width",
            "in": "query",
            "description": "Width",
            "required": false,
            "type": "integer"
          },
          {
            "name": "height",
            "in": "query",
            "description": "Height",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/grayscale": {
      "post": {
        "summary": "Convert an image to grayscale",
        "description": "Convert an image to grayscale",
        "tags": [
          "image"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "description": "Image file",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}